requirements.txt
RPi.GPIO
pyserial
ino
Flask
flask-socketio
eventlet